Rename latency -> accumulator
